requirements.txt
# Python Dependencies for Oura Personal Dashboard
# Install with: pip install -r requirements.txt

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Data analysis and visualization
# Uncomment if you want to analyze your data
# pandas>=2.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

streamlit
python-dotenv
requests

streamlit-authenticator

